requirements.txt
torch==1.0.1.post2
pyzmq==18.1.0
numpy==1.17.2
transformers==2.2.1